Add recharts.js
