+++ src/containers/TrackitList.js
@@ -1,12 +1,11 @@
 import React, { useState } from 'react';
 import { useSelector } from 'react-redux';
-// import { compact } from 'lodash';
 import Measurement from '../components/Measurement';
 import Nav from '../components/Nav';
 import Footer from '../components/Footer';
 import 'react-circular-progressbar/dist/styles.css';
 import TrackItDate from '../components/TrackItDate';
-import { readableDate, groupMeasurementByCreatedAt } from '../helpers/utils';
+import { convertedData, groupMeasurementByCreatedAt } from '../helpers/utils';
 import style from '../style/trackit.module.css';
 
 const TrackitList = () => {
@@ -18,22 +17,8 @@ const TrackitList = () => {
   const myData = measurements.measurements;
   const myMeasure = measures.measures;
 
-  const convertedData = myData.map((data) => {
-    const newData = {};
-    newData.value = data.value;
-    newData.user_id = data.user_id;
-    newData.measure_id = data.measure_id;
-    newData.created_at = readableDate(data.created_at);
-    const title = myMeasure
-      .filter((measure) => (data.measure_id === measure.id ? measure.title : ''));
-    newData.title = title[0].title;
-    return (newData);
-  });
-
-  console.log(convertedData);
-  const result = groupMeasurementByCreatedAt(convertedData, 'created_at');
-  // const result2 = groupMeasurementByMeasureId(convertedData, 'measure_id');
-  // console.log(result2);
+  // console.log(convertedData(myData, myMeasure));
+  const result = groupMeasurementByCreatedAt(convertedData(myData, myMeasure), 'created_at');
 
   const dateAndDetails = Object.keys(result);
   // console.log(dateAndDetails);
